Add api and router
- Move routing into it's own script
- Move emailing and file upload logic into their own apis
